UGENE-3335: gui test
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@7623 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Core/src/tasks/SequenceWalkerTask.cpp
@@ -65,7 +65,7 @@ QList<SequenceWalkerSubtask*> SequenceWalkerTask::prepareSubtasks() {
         config.range = wholeSeqReg.intersect(config.range);
     }
 
-    if (config.walkCircular && config.range.length == config.seqSize) {
+    if (config.walkCircular && static_cast<quint64>(config.range.length) == config.seqSize) {
         tempBuffer.clear();
         tempBuffer.append(QByteArray(config.seq, config.seqSize));
         tempBuffer.append(QByteArray(config.seq).left(config.walkCircularDistance * (config.aminoTrans == NULL ? 1 : 3)));
